requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
chromadb
sentence-transformers
httpx
yt-dlp
youtube-transcript-api
trafilatura
PyMuPDF
gitpython
openai-whisper
google-auth
google-auth-oauthlib
google-api-python-client
python-multipart
websockets
aiofiles
